requirements.txt
Kivy
Pillow
exifread
